fix: no return in calculatePlayerStatistics
Move the call of the function to upper component because the return value is used there.
